rename jsx to js
